pyproject.toml
[project]
name = "getsongsweb-api"
version = "1.0.0"
description = ""
authors = [
    {name = "Lucas Henrique",email = "lucash.rocha@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "meilisearch (>=0.38.0,<0.39.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "reportlab (>=4.4.5,<5.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "sentry-sdk[django] (>=2.46.0,<3.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "psycopg[binary] (>=3.2.13,<4.0.0)",
    "whitenoise (>=6.11.0,<7.0.0)"
]


[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.11.0,<26.0.0)"
]
